refactor: use decimal instead of integer

+++ migrations/Version20260124110742.php
@@ -20,9 +20,9 @@ final class Version20260124110742 extends AbstractMigration
             CREATE TABLE sensor_data (
               id INT AUTO_INCREMENT NOT NULL,
               node_uuid VARCHAR(36) NOT NULL,
-              temperature INT NOT NULL,
-              humidity INT NOT NULL,
-              pressure INT NOT NULL,
+              temperature DECIMAL(4, 1) NOT NULL,
+              humidity DECIMAL(4, 1) NOT NULL,
+              pressure DECIMAL(5, 1) NOT NULL,
               carbon_dioxide INT DEFAULT NULL,
               measured_at DATETIME NOT NULL,
               created_at DATETIME NOT NULL,
